Refactor contact modal/form
