+++ src/components/common/ContactModal/contact-form.scss
@@ -11,15 +11,13 @@
   .smilies-logo {
     margin-top: 3%;
   }
+}
 
-  .close-button {
-    font-size: 1.1rem;
-    color: $black;
-    background: none;
-    border: none;
-    margin-left: -90%;
-    margin-top: 2%;
-  }
+.close-button {
+  background: none;
+  border: none;
+  padding-left: 10px;
+  padding-top: 10px;
 }
 
 .contact-form {
@@ -32,11 +30,11 @@
   }
 
   .input-text {
-    padding-left: 30rem;
+    width: 300px;
   }
 
   .input-text-large {
-    padding-top: 5rem;
+    height: 100px;
   }
 }
