+++ src/styles/styles.scss
@@ -9,7 +9,7 @@
 @import '../components/common/Home/EmptyFeed/empty-feed.scss';
 @import '../components/common/Home/ChatList/MessagePreview/message-preview.scss';
 @import '../components/common/Home/ChatList/chat-list.scss';
-@import '../components/common/ContactForm/contact-form.scss';
+@import '../components/common/ContactModal/contact-form.scss';
 @import '../components/common/Home/home.scss';
 
 @import '../pages/LoginPage/login-page.scss';
